ajout cas vibratios

+++ src/c/mobile_mean_accel.h
@@ -3,6 +3,6 @@
 #define MOBILE_MEAN_ACCEL
   
   #include <pebble.h>
-  void mobile_mean_accel(AccelData *, uint32_t , uint32_t *);
+  void mobile_mean_accel(AccelData *, uint32_t , int16_t *);
 
 #endif
